requirements.txt
ffmpeg_python==0.2.0
matplotlib==3.7.2
moviepy==1.0.3
numpy==1.24.1
opencv_python==4.7.0.68
pafy==0.5.5
Pillow==10.0.0
scikit_learn==1.3.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
youtube_dl==2021.12.17
